Likes ,cards and chats UI

+++ thrifted/database/migrations/0001_01_01_000000_create_users_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->unsignedBigInteger('phone')->nullable();
             $table->boolean('delivery')->default(false);
             $table->boolean('payment_on_hand')->default(false);//need migration
-            $table->unsignedInteger('rating')->default(0);
+            $table->float('rating')->default(0);
             $table->timestamp('email_verified_at')->nullable();
             $table->string('password');
             $table->string('picture')->nullable();
